Low-level driver loader

Most users should use adbc_driver(); however, this function may be used to allow other libraries (e.g., GDAL) to access the driver loader.

Usage

adbc_driver_load(
  x,
  entrypoint,
  version,
  driver,
  error,
  load_flags = adbc_load_flags(),
  additional_search_path_list = NULL
)

Arguments

x, entrypoint

An ADBC driver may be defined either as an init function or as an identifier with an entrypoint name. A driver init func must be an external pointer to a DL_FUNC with the type AdbcDriverInitFunc specified in the adbc.h header.

version

The version number corresponding to the driver supplied

driver

An external pointer to an AdbcDriver

error

An external pointer to an AdbcError or NULL

load_flags

Integer flags generated by adbc_load_flags()

additional_search_path_list

A path list of additional locations to search for driver manifests

Value

An integer ADBC status code