Create ADBC drivers

Creates the R object representation of an ADBC driver, which consists of a name and an initializer function with an optional subclass to control finer-grained behaviour at the R level.

Usage

adbc_driver_void()

adbc_driver(
  x,
  entrypoint = NULL,
  ...,
  load_flags = adbc_load_flags(),
  subclass = character()
)

Arguments

x, entrypoint

An ADBC driver may be defined either as an init function or as an identifier with an entrypoint name. A driver init func must be an external pointer to a DL_FUNC with the type AdbcDriverInitFunc specified in the adbc.h header.

...

Further key/value parameters to store with the (R-level) driver object.

load_flags

Integer flags generated by adbc_load_flags()

subclass

An optional subclass for finer-grained control of behaviour at the R level.

Value

An object of class 'adbc_driver'

Examples

adbc_driver_void()
#> <adbc_driver_void> List of 4
#>  $ load_flags      : int 15
#>  $ driver_init_func:Class 'adbc_driver_init_func' <externalptr> 
#>  $ .child_count    : int 0
#>  $ version         : int 1001000